% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertGroupby.R
\name{insertGroupby}
\alias{insertGroupby}
\title{Insert dplyr::group_by(.}
\usage{
insertGroupby()
}
\description{
Call this function as an addin to insert \code{ dplyr::group_by( } at the cursor position.
}
