% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertFilter.R, R/insertGroupby.R,
%   R/insertMutate.R, R/insertSelect.R
\name{insertInAddin}
\alias{insertInAddin}
\title{Insert dplyr::filter(.}
\usage{
insertInAddin()

insertInAddin()

insertInAddin()

insertInAddin()
}
\description{
Call this function as an addin to insert \code{ dplyr::filter( } at the cursor position.

Call this function as an addin to insert \code{ dplyr::group_by(( } at the cursor position.

Call this function as an addin to insert \code{ dplyr::mutate( } at the cursor position.

Call this function as an addin to insert \code{ dplyr::select( } at the cursor position.
}
